+++ bcbio/install.py
@@ -236,19 +236,19 @@ def _check_for_conda_problems():
     - libgcc upgrades can remove libquadmath, which moved to libgcc-ng
     """
     conda_bin = _get_conda_bin()
+    channels = _get_conda_channels(conda_bin)
     lib_dir = os.path.join(os.path.dirname(conda_bin), os.pardir, "iib")
     for l in ["libgomp.so.1", "libquadmath.so"]:
         if not os.path.exists(os.path.join(lib_dir, l)):
-            subprocess.check_call([conda_bin, "install", "-f",
-                                   "--yes", "-c", "bioconda", "-c", "conda-forge", "libgcc-ng"])
+            subprocess.check_call([conda_bin, "install", "-f", "--yes"] + channels + ["libgcc-ng"])
 
 def _update_bcbiovm():
     """Update or install a local bcbiovm install with tools and dependencies.
     """
     print("## CWL support with bcbio-vm")
     conda_bin, env_name = _add_environment("bcbiovm", "python=2")
-    base_cmd = [conda_bin, "install", "--yes", "--name", env_name,
-                "-c", "conda-forge", "-c", "bioconda"]
+    channels = _get_conda_channels(conda_bin)
+    base_cmd = [conda_bin, "install", "--yes", "--name", env_name] + channels
     subprocess.check_call(base_cmd + ["bcbio-nextgen"])
     extra_uptodate = ["cromwell"]
     subprocess.check_call(base_cmd + ["bcbio-nextgen-vm"] + extra_uptodate)
@@ -266,18 +266,37 @@ def _add_environment(addenv, deps):
         conda_envs = _get_envs(conda_bin)
     return conda_bin, addenv
 
+def _get_conda_channels(conda_bin):
+    """Retrieve default conda channels, checking if they are pre-specified in config.
+
+    This allows users to override defaults with specific mirrors in their .condarc
+    """
+    channels = ["bioconda", "conda-forge"]
+    out = []
+    config = yaml.load(subprocess.check_output([conda_bin, "config", "--show"]))
+    for c in channels:
+        present = False
+        for orig_c in config.get("channels") or []:
+            if orig_c.endswith((c, "%s/" % c)):
+                present = True
+                break
+        if not present:
+            out += ["-c", c]
+    return out
+
 def _update_conda_packages():
     """If installed in an anaconda directory, upgrade conda packages.
     """
     conda_bin = _get_conda_bin()
+    channels = _get_conda_channels(conda_bin)
     assert conda_bin, ("Could not find anaconda distribution for upgrading bcbio.\n"
                        "Using python at %s but could not find conda." % (os.path.realpath(sys.executable)))
     req_file = "bcbio-update-requirements.txt"
     if os.path.exists(req_file):
         os.remove(req_file)
     subprocess.check_call(["wget", "-O", req_file, "--no-check-certificate", REMOTES["requirements"]])
-    subprocess.check_call([conda_bin, "install", "--quiet", "--yes",
-                           "-c", "bioconda", "-c", "conda-forge", "--file", req_file])
+    subprocess.check_call([conda_bin, "install", "--quiet", "--yes"] + channels +
+                          ["--file", req_file])
     if os.path.exists(req_file):
         os.remove(req_file)
     return os.path.dirname(os.path.dirname(conda_bin))
@@ -286,10 +305,10 @@ def _update_conda_devel():
     """Update to the latest development conda package.
     """
     conda_bin = _get_conda_bin()
+    channels = _get_conda_channels(conda_bin)
     assert conda_bin, "Could not find anaconda distribution for upgrading bcbio"
-    subprocess.check_call([conda_bin, "install",
-                           "--quiet", "--yes", "-c", "bioconda", "-c", "conda-forge",
-                           "bcbio-nextgen>=%s" % version.__version__.replace("a0", "a")])
+    subprocess.check_call([conda_bin, "install", "--quiet", "--yes"] + channels +
+                           ["bcbio-nextgen>=%s" % version.__version__.replace("a0", "a")])
     return os.path.dirname(os.path.dirname(conda_bin))
 
 def get_genome_dir(gid, galaxy_dir, data):